chore: rename  CSS modules to match associated .tsx files
