+++ app/about/team/page.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { teamMembers } from "@/app/data";
 import TeamCard from "@/app/components/teamCard/TeamCard";
-import styles from "./our-team.module.css";
+import styles from "./page.module.css";
 
 const Team = () => {
   return (
